requirements.txt
gunicorn
dj-database-url
python-dotenv
whitenoise
psycopg2-binary
